requirements.txt
selenium>=4.27.1,<4.28.0
pytest>=8.3.3,<8.4.0
pytest-xdist>=3.6.1,<3.4.0
allure-pytest>=2.13.5,<2.14.0
requests>=2.32.3,<2.33.0
python-dotenv>=1.0.1,<1.1.0